+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `disallowDefaultNetwork`
   - `requireInstanceCmek`
   - `requireNoPublicIp`
+- Projects policies [#41](https://github.com/losisin/gcp-pac/pull/41)
+  - `requireApiKeyRestrictions`
 
 ## [1.7.0] - 2024-01-22